refactor(oauth): split session module for GH727
