+++ specs/GH727/product.md
@@ -6,10 +6,9 @@ GH-727 / #727
 
 ## 用户问题
 
-`origin/main@e7fd7a121a69` 仍有 29 个 tracked Rust 文件超过 U-16 的 800 行硬上限。当前最大文件是
-`src/core/router/tests/strategy_impl_tests.rs`，它是一个 880 行 test-only suite，把 routing context
-construction、weighted random、least busy、lowest usage、lowest latency、lowest priority、rate-limit-aware、
-round-robin 和 strategy consistency tests 全部放在一个文件里。
+`origin/main@38b3140aeeca` 仍有 28 个 tracked Rust 文件超过 U-16 的 800 行硬上限。当前最大文件是
+`src/auth/oauth/session.rs`，它是一个 869 行 runtime session module，把 OAuth session model、
+session store trait/error、in-memory store、Redis store 和 inline tests 全部放在一个文件里。
 
 本轮目标继续执行完整的大文件解耦计划：每个 PR 仍然小而可审，但所有 tranche 都必须服从同一套
 架构边界，避免制造新的耦合、重导出混乱或行为漂移。
@@ -35,42 +34,42 @@ round-robin 和 strategy consistency tests 全部放在一个文件里。
 
 ## 本 tranche 目标
 
-- 拆分 `src/core/router/tests/strategy_impl_tests.rs`，它当前 880 行，是 #727 当前最大的 test-only suite。
-- 保留 `src/core/router/tests/mod.rs` 的 `mod strategy_impl_tests;` 挂载方式不变。
-- 将 root `strategy_impl_tests.rs` 缩小为 shared imports、provider/deployment helpers 和 strategy-domain child module declarations。
-- 按策略域移动原测试断言到 `src/core/router/tests/strategy_impl_tests/*.rs`：
-  context builder、weighted random、least busy、lowest usage、lowest latency、lowest priority、rate-limit-aware、round-robin、integration consistency。
-- 保持非重复 strategy coverage、async behavior、atomic counter behavior 和 focused test command 覆盖不变。
-- 已按 review 删除一个只重复首轮 round-robin 覆盖的 test；`test_round_robin_cycles_through_candidates`
-  继续覆盖相同首轮顺序和下一轮 wrap 行为。
+- 拆分 `src/auth/oauth/session.rs`，它当前 869 行，是 #727 当前最大的 runtime module。
+- 保留 `src/auth/oauth/mod.rs` 的 `pub mod session;` 和公开导入路径不变。
+- 将 root `session.rs` 缩小为 facade，只声明 child modules 并 re-export `OAuthSession`、`SessionStore`、
+  `SessionError`、`InMemorySessionStore` 和 feature-gated `RedisSessionStore`。
+- 按职责移动实现到 `src/auth/oauth/session/*.rs`：
+  `model.rs`、`store.rs`、`memory_store.rs`、`redis_store.rs`、`tests.rs`。
+- 保持 session expiration、token update、state deletion、user session indexing、Redis key/TTL behavior 和 tests 覆盖不变。
 - 所有新增或修改后的 Rust 文件低于 800 行。
 
 ## 非目标
 
-- 不修改 `src/core/router/strategy_impl.rs`、deployment model 或 router runtime selection behavior。
-- 不改变任何 routing strategy scoring、selection、counter、limit、latency 或 priority assertion。
-- 不在本 PR 中处理其余 28 个大文件。
+- 不修改 OAuth handlers、middleware、client、provider discovery 或 user creation logic。
+- 不改变 session serialization、expiration cleanup、state CSRF deletion、Redis key prefix 或 TTL semantics。
+- 不在本 PR 中处理其余 27 个大文件。
 - 不关闭 #727。
 
 ## Behavior Invariants
 
-1. `src/core/router/tests/mod.rs` continues to mount the suite as `core::router::tests::strategy_impl_tests`.
-2. Shared provider and deployment helpers remain test-only helpers in the strategy test root.
-3. Existing weighted random, least busy, lowest usage, lowest latency, lowest priority, rate-limit-aware, round-robin, and consistency coverage remains discoverable, except the review-confirmed duplicate round-robin first-cycle test is removed.
-4. Round-robin tests keep the same `DashMap<String, AtomicUsize>` counter behavior.
-5. Every touched Rust file must be below U-16's 800-line ceiling.
-6. `cargo test core::router::tests::strategy_impl_tests --lib --all-features` must pass.
+1. `auth::oauth::session::OAuthSession`, `SessionStore`, `SessionError`, `InMemorySessionStore`, and Redis-gated `RedisSessionStore` remain available from the original module path.
+2. `OAuthSession` fields, serde attributes, builder methods, expiration checks, `touch`, `extend`, and `update_token` semantics remain unchanged.
+3. `SessionStore` trait method signatures remain unchanged.
+4. In-memory store session/state cleanup semantics remain unchanged.
+5. Redis store key format, set membership, TTL, serialization, and `get_del` state behavior remain unchanged.
+6. Every touched Rust file must be below U-16's 800-line ceiling.
+7. `cargo test auth::oauth::session --lib --all-features` must pass.
 
 ## 验收标准
 
-- [ ] `src/core/router/tests/strategy_impl_tests.rs` keeps only shared helpers and child module declarations。
-- [ ] Original strategy implementation tests move under `src/core/router/tests/strategy_impl_tests/*.rs` without assertion changes。
-- [ ] Strategy coverage remains present for routing contexts, weighted random, least busy, lowest usage, latency, priority, rate limits, round robin, and consistency。
-- [ ] All touched Router strategy test files are below U-16's 800-line ceiling。
-- [ ] Focused Router strategy test suite 通过。
+- [ ] `src/auth/oauth/session.rs` is a small facade with module declarations and public re-exports。
+- [ ] Session model, store trait/error, in-memory store, Redis store, and tests move under `src/auth/oauth/session/*.rs` without API path changes。
+- [ ] OAuth session expiration, state, cleanup, user-session, Redis, and serialization coverage remains present。
+- [ ] All touched OAuth session files are below U-16's 800-line ceiling。
+- [ ] Focused OAuth session test suite 通过。
 - [ ] `cargo fmt --all -- --check`、`cargo check --lib --all-features`、`cargo check --all-features --locked` 和 `cargo check` 通过。
 - [ ] PR body 明确该 PR 是 #727 的 partial tranche，使用 `Refs #727`，不自动关闭 tracker issue。
 
 ## 发布说明
 
-No runtime behavior change. This is a Router strategy test-suite split for U-16 compliance.
+No runtime behavior change. This is an OAuth session module split for U-16 compliance.
